Add add to calendar functionality
This is a really crappy implementation of an add to calendar feature
It looks for a date value in the SCHEDULE markdown file (by convention)
and it shows a third call to action button with a Google calendar link.

This is mostly experimental and should be looked as something we can do
in the future, maybe we can also explore the Meetup API as we've been
talking about.