Show Schedule from the Meetup API
We stop using the schedule.md strategy in favor of getting
that information from the Meetup API. From now on we'll use
that as our source of truth.

First we use a JSONp strategy to bypass some of the meetup API's
constraints, I think there's a cap on the number of requests that
one can have daily to the meetup API but that constrait is pushed to
the visitor.

Then, we build our Schedule with the data atribute from the
description property of the returned data. We still use `snarkdown`
to parse the response even though Meetup API's response is just an HTML
string, that's because we'd tend to write the Meetup description in a
markdown-like style.

Then, we build our Calendar link, this is going to happen on a further
commit.

And finally we build our Schedule title, it shows the month in which
the Meetup is happening and the Google maps link that's being built
with the Meetup API's lat and long part of the response.